added login logout and restriction features

+++ base/views.py
@@ -1,18 +1,19 @@
 from django.shortcuts import render, redirect
 from django.contrib.auth.models import User
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.http import HttpResponse
 from django.contrib import messages
 from django.db.models import Q
 from . models import Room, Topic
 from . forms import RoomForm
 
-# rooms = [
-#     {'id':1, 'name':'Lets learn'},
-#     {'id':2, 'name':'Lets build'},
-#     {'id':3, 'name':'Lets debug'},
-#     {'id':4, 'name':'Lets deploy'},
-# ]
 
 def loginPage(request):
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
     if request.method == 'POST':
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -21,9 +22,19 @@ def loginPage(request):
             user = User.objects.get(username=username)
         except:
             messages.error(request,'User does not exists')
+        user  = authenticate(request, username=username, password=password)
+        if user is not None:
+            login(request,user)
+            return redirect('home')
+        else:
+            messages.error(request,'Username or Passsword does not exist')
     context = {}
     return render(request, 'base/login_register.html', context)
 
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
 
 
 def home(request):
@@ -40,11 +51,14 @@ def home(request):
     context = {'rooms':rooms, 'topics':topics, 'room_count':room_count}
     return render(request,'base/home.html', context)
 
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room':room}
     return render(request,'base/room.html', context)
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     form  = RoomForm()
     if request.method == 'POST':
@@ -55,9 +69,15 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
+
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -66,8 +86,12 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request,'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here')
     if request.method == 'POST':
         room.delete()
         return redirect('home')
